refactor(runtime): split orchestrator/tests — extract end_turn + empty FinalAnswer
Extract the 6 reply-tool-contract tests into a new
`tests/end_turn.rs`:

- `end_turn_rejected_when_reply_tool_exists_but_not_called`
- `end_turn_accepted_without_reply_tool_in_cli_mode`
- `end_turn_accepted_after_reply_tool_called`
- `end_turn_accepted_after_react_tool_called`
- `empty_final_answer_not_persisted_and_retries`
- `empty_final_answer_not_persisted_in_run_turn`

`MockReplyExtTool` moves with them (only used by the empty-FinalAnswer
tests). `MockExtTool` stays in `tests/mod.rs` since it's also used by
the attachment-collection tests still living there; sibling modules
reach it via `super::MockExtTool`.

cargo test -p assistant-runtime --lib reports 220 passed (same as
before); `orchestrator::tests::end_turn::*` subtotal is 6.

Second slice in the stack — more sections to follow (multimodal,
sanitize, bus/worker, streaming, misc).
